services/ccxtData/requirements.txt
ccxt==4.2.15
python-dotenv==1.0.0
aiohttp==3.9.3
asyncio==3.4.3 